backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
alembic==1.18.1
python-multipart==0.0.21
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
